fix: separate input system
