+++ cradle/linux/src/main.rs
@@ -12,6 +12,7 @@ use std::os::unix::io::{AsRawFd, RawFd};
 mod udev;
 mod epoll;
 mod kernel_input;
+mod input;
 mod userlib;
 
 pub struct PlatformAssetLoader { basedir: PathBuf }
@@ -77,13 +78,11 @@ impl peridot::NativeLinker for NativeLink
 }
 
 #[allow(dead_code)]
-struct X11
-{
+pub struct X11 {
     con: xcb::Connection, wm_protocols: xcb::Atom, wm_delete_window: xcb::Atom, vis: xcb::Visualid,
     mainwnd_id: xcb::Window
 }
-impl X11
-{
+impl X11 {
     fn init() -> Self {
         let (con, screen_index) = xcb::Connection::connect(None).expect("Connecting with xcb");
         let s0 = con.get_setup().roots().nth(screen_index as _).expect("No screen");
@@ -140,7 +139,7 @@ pub struct GameDriver {
     usercode: userlib::Game<NativeLink>
 }
 impl GameDriver {
-    fn new(wh: WindowHandler) -> Self {
+    fn new(wh: WindowHandler, x11: &std::rc::Rc<X11>) -> Self {
         let nl = NativeLink {
             al: PlatformAssetLoader::new(),
             wh
@@ -150,6 +149,7 @@ impl GameDriver {
             nl, userlib::Game::<NativeLink>::requested_features()
         );
         let usercode = userlib::Game::init(&mut engine);
+        engine.input_mut().set_nativelink(Box::new(input::InputNativeLink::new(x11)));
         engine.postinit();
 
         GameDriver { engine, usercode }
@@ -158,193 +158,25 @@ impl GameDriver {
     fn update(&mut self) { self.engine.do_update(&mut self.usercode); }
 }
 
-fn lookup_device_name(d: &udev::Device) -> Option<String> {
-    match d.property_value(unsafe { std::ffi::CStr::from_bytes_with_nul_unchecked(b"NAME\0") }) {
-        Some(n) => Some(String::from(n.to_str().expect("decoding failed"))),
-        None => d.parent().as_ref().and_then(lookup_device_name)
-    }
-}
-#[allow(dead_code)]
-fn diag_device(d: &udev::Device) {
-    for p in d.iter_properties() {
-        let name = p.name().expect("no name?").to_str().expect("decoding failed");
-        if let Some(v) = p.value() {
-            println!("  * {} = {}", name, v.to_str().expect("decoding failed value"));
-        } else {
-            println!("  * {}", name);
-        }
-    }
-    if let Some(p) = d.parent() {
-        println!(
-            "    * name = {}",
-            p.property_value(unsafe { std::ffi::CStr::from_bytes_with_nul_unchecked(b"NAME\0") })
-                .map_or("<none>", |s| s.to_str().expect("decoding failed"))
-        );
-        println!(
-            "    * devtype = {}",
-            p.devtype().map_or("<none>", |s| s.to_str().expect("decoding failed"))
-        );
-        println!(
-            "    * devnode = {}",
-            p.devnode().map_or("<none>", |s| s.to_str().expect("decoding failed"))
-        );
-        println!("    * initialized ? {}", p.is_initialized());
-        for p in p.iter_properties() {
-            let name = p.name().expect("no name?").to_str().expect("decoding failed");
-            if let Some(v) = p.value() {
-                println!("    * {} = {}", name, v.to_str().expect("decoding failed value"));
-            } else {
-                println!("    * {}", name);
-            }
-        }
-    }
-}
-
-use std::collections::{BTreeMap, BTreeSet, HashMap};
-
-pub struct EventDevice { fd: libc::c_int, is_mouse: bool }
-impl EventDevice {
-    pub fn open(path: &std::ffi::CStr, is_mouse: bool) -> std::io::Result<Self> {
-        let fp = unsafe { libc::open(path.as_ptr(), libc::O_RDONLY) };
-        if fp < 0 { Err(std::io::Error::last_os_error()) } else { Ok(Self { fd: fp, is_mouse }) }
-    }
-
-    pub fn read(&self) -> std::io::Result<kernel_input::InputEvent> {
-        let mut ev = std::mem::MaybeUninit::uninit();
-        let r = unsafe { libc::read(self.fd, ev.as_mut_ptr() as _, std::mem::size_of::<kernel_input::InputEvent>()) };
-        if r < 0 { Err(std::io::Error::last_os_error()) } else { Ok(unsafe { ev.assume_init() }) }
-    }
-}
-impl Drop for EventDevice {
-    fn drop(&mut self) { unsafe { libc::close(self.fd); } }
-}
-
-pub struct EventDeviceManager {
-    pub epoll_id_resv: u64,
-    pub devices_by_id: BTreeMap<u64, EventDevice>,
-    pub id_free_list: BTreeSet<u64>,
-    pub device_id_by_node: HashMap<String, u64>
-}
-impl EventDeviceManager {
-    pub fn new(
-        initial_devices: impl Iterator<Item = (String, EventDevice)>, epoll_id_resv: u64, epoll: &epoll::Epoll
-    ) -> Self {
-        let mut devices_by_id = BTreeMap::new();
-        let mut device_id_by_node = HashMap::with_capacity(initial_devices.size_hint().0);
-        for (n, (node, d)) in initial_devices.enumerate() {
-            epoll.add_fd(d.fd, libc::EPOLLIN as _, n as u64 + epoll_id_resv)
-                .expect("Failed to register initial device fd to epoll");
-            devices_by_id.insert(n as u64, d);
-            device_id_by_node.insert(node, n as u64);
-        }
-
-        EventDeviceManager {
-            epoll_id_resv,
-            devices_by_id,
-            id_free_list: BTreeSet::new(),
-            device_id_by_node
-        }
-    }
-    pub fn len(&self) -> usize { self.devices_by_id.len() }
-
-    /// Returns assigned id(in EventDeviceManager)
-    pub fn add(&mut self, node: String, device: EventDevice, epoll: &epoll::Epoll) -> u64 {
-        let id = self.id_free_list.pop_first().unwrap_or_else(|| self.devices_by_id.len() as u64);
-        epoll.add_fd(device.fd, libc::EPOLLIN as _, id + self.epoll_id_resv)
-            .expect("Failed to register device fd to epoll");
-        self.devices_by_id.insert(id, device);
-        self.device_id_by_node.insert(node, id);
-        
-        id
-    }
-    pub fn remove(&mut self, id: u64, epoll: &epoll::Epoll) {
-        if let Some(d) = self.devices_by_id.remove(&id) {
-            self.id_free_list.insert(id);
-            epoll.remove_fd(d.fd).expect("Failed to unregister device fd from epoll");
-        }
-    }
-
-    pub fn lookup_id_by_node(&self, node: &str) -> Option<u64> { self.device_id_by_node.get(node).copied() }
-    pub fn translate_epoll_value(&self, epoll_value: u64) -> u64 { epoll_value - self.epoll_id_resv }
-    pub fn get_device(&self, id: u64) -> Option<&EventDevice> { self.devices_by_id.get(&id) }
-}
-
-pub struct InputNativeLink {
-    ws_ref: std::rc::Weak<X11>
-}
-impl peridot::NativeInput for InputNativeLink {
-    fn get_pointer_position(&self, _index: u32) -> Option<(f32, f32)> {
-        if let Some(ws) = self.ws_ref.upgrade() {
-            let ck = xcb::query_pointer(&ws.con, ws.mainwnd_id);
-            ws.con.flush();
-            let ptrinfo = ck.get_reply().expect("Failed to query pointer to xcb");
-            if ptrinfo.same_screen() {
-                // Note: なぜかLinux/XCBでも5.0だけずれるんですけど！！
-                Some((ptrinfo.win_x() as _, ptrinfo.win_y() as f32 - 5.0))
-            } else {
-                debug!("Fixme: Handle same_screen = false");
-                None
-            }
-        } else {
-            None
-        }
-    }
-}
-
 fn main() {
     env_logger::init();
     let x11 = std::rc::Rc::new(X11::init());
 
-    let udev = udev::Context::new().expect("Failed to initialize udev");
-    let enumerator = udev::Enumerate::new(&udev).expect("Failed to create udev Enumerator");
-    let target_subsystem_name = std::ffi::CString::new("input").expect("encoding failed");
-    enumerator.add_match_subsystem(&target_subsystem_name).expect("Failed to set subsystem filter");
-    enumerator.add_match_is_initialized().expect("FAiled to set initialized filter");
-    enumerator.scan_devices().expect("Failed to scan devices");
-    let id_input_mouse = unsafe { std::ffi::CStr::from_bytes_with_nul_unchecked(b"ID_INPUT_MOUSE\0") };
-    let event_device_regex = regex::Regex::new("event[0-9]+$").expect("invalid regex");
-    let initial_devices = enumerator.iter().filter_map(|e| {
-        let syspath = e.name().expect("no name?");
-        let device = udev::Device::from_syspath(&udev, syspath).expect("Failed to create udev Device");
-        let devnode_c = match device.devnode() {
-            Some(s) => s,
-            None => return None
-        };
-        let devnode = devnode_c.to_str().expect("decoding failed");
-        if !event_device_regex.is_match(devnode) { return None; }
-        let device_name = lookup_device_name(&device).unwrap_or_else(|| String::from("<unknown device name>"));
-        // diag_device(&device);
-        let is_mouse = device.property_value(id_input_mouse).map_or(false, |s| s.to_str() == Ok("1"));
-
-        info!("Registering Input Device: {} ({})", devnode, device_name);
-        
-        Some((String::from(devnode), EventDevice::open(devnode_c, is_mouse).expect("Failed to open event device")))
-    });
-    let monitor = udev::Monitor::from_netlink(
-        &udev, unsafe { std::ffi::CStr::from_bytes_with_nul_unchecked(b"udev\0") }
-    ).expect("Failed to create udev monitor");
-    monitor.filter_add_match_subsystem_devtype(
-        unsafe { std::ffi::CStr::from_bytes_with_nul_unchecked(b"input\0") },
-        None
-    ).expect("Failed to set monitor subsystem filter");
-    monitor.filter_update().expect("Failed to update monitor filter");
-    monitor.enable_receiving().expect("Failed to set monitor receiving enable");
-    let monitor_fd = monitor.fd().expect("Failed to retrieve udev monitor fd");
-
-    let mut gd = GameDriver::new(WindowHandler { dp: x11.con.get_raw_conn(), vis: x11.vis, wid: x11.mainwnd_id });
-    gd.engine.input_mut().set_nativelink(Box::new(InputNativeLink { ws_ref: std::rc::Rc::downgrade(&x11) }));
+    let mut gd = GameDriver::new(
+        WindowHandler { dp: x11.con.get_raw_conn(), vis: x11.vis, wid: x11.mainwnd_id },
+        &x11
+    );
 
     let ep = epoll::Epoll::new().expect("Failed to create epoll interface");
     ep.add_fd(x11.fd(), libc::EPOLLIN as _, 0).expect("Failed to add x11 fd");
-    ep.add_fd(monitor_fd, libc::EPOLLIN as _, 1).expect("Failed to add udev monitor fd");
-    let mut devmgr = EventDeviceManager::new(initial_devices, 2, &ep);
+    let mut input = input::InputSystem::new(&ep, 1, 2);
 
     x11.show();
-    let mut events = vec![unsafe { std::mem::MaybeUninit::zeroed().assume_init() }; 2 + devmgr.len()];
+    let mut events = vec![unsafe { std::mem::MaybeUninit::zeroed().assume_init() }; 2 + input.managed_devices_count()];
     'app: loop {
-        if events.len() != 2 + devmgr.len() {
+        if events.len() != 2 + input.managed_devices_count() {
             // resize
-            events.resize(2 + devmgr.len(), unsafe { std::mem::MaybeUninit::zeroed().assume_init() });
+            events.resize(2 + input.managed_devices_count(), unsafe { std::mem::MaybeUninit::zeroed().assume_init() });
         }
 
         let count = ep.wait(&mut events, Some(1)).expect("Failed to waiting epoll");
@@ -355,223 +187,11 @@ fn main() {
             if e.u64 == 0 {
                 if !x11.process_all_events() { break 'app; }
             } else if e.u64 == 1 {
-                let device = monitor.receive_device().expect("no device received?");
-                let devnode_c = match device.devnode() {
-                    Some(s) => s,
-                    None => continue
-                };
-                let devnode = devnode_c.to_str().expect("decoding failed");
-                if !event_device_regex.is_match(devnode) { continue; }
-                let action = device.action().to_str().expect("action decoding failed");
-                let device_name = lookup_device_name(&device).unwrap_or_else(|| String::from("<unknown device name>"));
-                println!("Incoming Device Event: {}", action);
-                // diag_device(&device);
-
-                match action {
-                    "add" => {
-                        let is_mouse = device.property_value(id_input_mouse).map_or(false, |s| s.to_str() == Ok("1"));
-                        info!("Registering Input Device: {} ({})", devnode, device_name);
-                        devmgr.add(
-                            String::from(devnode),
-                            EventDevice::open(devnode_c, is_mouse).expect("Failed to open added device"),
-                            &ep
-                        );
-                    },
-                    "remove" => if let Some(id) = devmgr.lookup_id_by_node(devnode) {
-                        info!( "Unregistering Input Device: {} ({})", devnode, device_name);
-                        devmgr.remove(id, &ep);
-                    },
-                    a => debug!("Unknown device action: {:?}", a)
-                }
+                input.process_monitor_event(&ep);
             } else {
-                let (ev, is_mouse) = match devmgr.get_device(devmgr.translate_epoll_value(e.u64)) {
-                    Some(d) => match d.read() {
-                        Ok(ev) => (ev, d.is_mouse),
-                        Err(e) => {
-                            error!("Failed to read from device: {:?}", e);
-                            continue;
-                        }
-                    },
-                    None => {
-                        let evnum = unsafe { std::ptr::read_unaligned(&e.u64) };
-                        error!("device not found? ev={}", evnum);
-                        continue;
-                    }
-                };
-                
-                if ev.type_ == kernel_input::EventType::Synchronize as u16 {
-                    // ignore dataframe sync
-                    if ev.code == 0 { continue; }
-
-                    debug!("syn event: code={}, value={}", ev.code, ev.value);
-                } else if ev.type_ == kernel_input::EventType::Key as u16 {
-                    let is_press = match ev.value {
-                        0 => false,
-                        1 => true,
-                        _ => continue
-                    };
-
-                    if let Some(b) = map_key_button(ev.code) {
-                        gd.engine.input().dispatch_button_event(b, is_press);
-                    } else if let Some(a) = map_analog_key_emulation(ev.code) {
-                        gd.engine.input().dispatch_analog_event(a, if is_press { 1.0 } else { 0.0 }, true);
-                    } else {
-                        debug!("key event: code={}", ev.code);
-                    }
-                } else if ev.type_ == kernel_input::EventType::Relative as u16 {
-                    if let Some(b) = if is_mouse { map_mouse_input_rel(ev.code) } else { map_input_rel(ev.code) } {
-                        gd.engine.input().dispatch_analog_event(b, ev.value as _, false);
-                    } else {
-                        debug!("relative event: code={}, value={}", ev.code, ev.value);
-                    }
-                } else if ev.type_ == kernel_input::EventType::Absolute as u16 {
-                    // ignore misc event from mouse, bitmask of pressed buttons
-                    if is_mouse && ev.code == kernel_input::AbsoluteAxes::Misc as u16 { continue; }
-
-                    // Todo: hat -> pov button translation
-                    if let Some(b) = if is_mouse { map_mouse_input_abs(ev.code) } else { map_input_abs(ev.code) } {
-                        gd.engine.input().dispatch_analog_event(b, ev.value as _, true);
-                    } else {
-                        debug!("absolute event: code={}, value={}", ev.code, ev.value);
-                    }
-                } else {
-                    // ignore keyscan misc
-                    if ev.type_ == kernel_input::EventType::Misc as u16 && ev.code == 4 { continue; }
-                    debug!("Other event: type={}, code={}, value={}", ev.type_, ev.code, ev.value);
-                }
+                input.process_device_event(gd.engine.input(), e.u64);
             }
         }
     }
     info!("Terminating Program...");
-}
-
-fn map_analog_key_emulation(key: u16) -> Option<peridot::NativeAnalogInput> {
-    if key == 0x138 { return Some(peridot::NativeAnalogInput::LeftTrigger); }
-    if key == 0x139 { return Some(peridot::NativeAnalogInput::RightTrigger); }
-
-    None
-}
-fn map_key_button(key: u16) -> Option<peridot::NativeButtonInput> {
-    use peridot::NativeButtonInput::*;
-    use peridot::NativeButtonInput::Character as C;
-
-    const KEYBOARD_MAP: &[peridot::NativeButtonInput] = &[
-        Esc,
-        C('1'), C('2'), C('3'), C('4'), C('5'), C('6'), C('7'), C('8'), C('9'), C('0'), C('-'), C('='),
-        Backspace,
-        C('\t'), C('Q'), C('W'), C('E'), C('R'), C('T'), C('Y'), C('U'), C('I'), C('O'), C('P'), C('{'), C('}'),
-        Enter, LeftControl,
-        C('A'), C('S'), C('D'), C('F'), C('G'), C('H'), C('J'), C('K'), C('L'), C(';'), C('\''), C('`'),
-        LeftShift, C('\\'),
-        C('Z'), C('X'), C('C'), C('V'), C('B'), C('N'), C('M'), C(','), C('.'), C('/'),
-        RightShift, C('*'), LeftAlt, C(' '),
-        CapsLock,
-        FunctionKey(1), FunctionKey(2),
-        FunctionKey(3), FunctionKey(4),
-        FunctionKey(5), FunctionKey(6),
-        FunctionKey(7), FunctionKey(8),
-        FunctionKey(9), FunctionKey(10)
-    ];
-    const GAMEPAD_MAP: &[Option<peridot::NativeButtonInput>] = &[
-        Some(ButtonA), Some(ButtonB), Some(ButtonC),
-        Some(ButtonX), Some(ButtonY), None/*ButtonZ*/,
-        Some(ButtonL), Some(ButtonR),
-        None, None,
-        Some(ButtonSelect), Some(ButtonStart), None,
-        Some(Stick(0)), Some(Stick(1))
-    ];
-
-    if (1 ..= 68).contains(&key) { return Some(KEYBOARD_MAP[key as usize - 1]); }
-    if key == 85 { return Some(peridot::NativeButtonInput::ZenkakuHankaku); }
-    if key == 87 { return Some(peridot::NativeButtonInput::FunctionKey(11)); }
-    if key == 88 { return Some(peridot::NativeButtonInput::FunctionKey(12)); }
-    if key == 97 { return Some(peridot::NativeButtonInput::RightControl); }
-    if key == 100 { return Some(peridot::NativeButtonInput::RightAlt); }
-    if key == 103 { return Some(peridot::NativeButtonInput::UpArrow); }
-    if key == 105 { return Some(peridot::NativeButtonInput::LeftArrow); }
-    if key == 106 { return Some(peridot::NativeButtonInput::RightArrow); }
-    if key == 108 { return Some(peridot::NativeButtonInput::DownArrow); }
-    if key == 125 { return Some(peridot::NativeButtonInput::LeftMeta); }
-    if key == 126 { return Some(peridot::NativeButtonInput::RightMeta); }
-    if (183 ..= 194).contains(&key) {
-        return Some(peridot::NativeButtonInput::FunctionKey(13 + (key - 183) as u8));
-    }
-    if (0x130 ..= 0x13e).contains(&key) { return GAMEPAD_MAP[key as usize - 0x130]; }
-    
-    if (kernel_input::Key::Left as u16 ..= kernel_input::Key::Task as u16).contains(&key) {
-        return Some(peridot::NativeButtonInput::Mouse((key - kernel_input::Key::Left as u16) as _));
-    }
-
-    debug!("key event: code={}", key);
-    None
-}
-fn map_mouse_input_rel(code: u16) -> Option<peridot::NativeAnalogInput> {
-    if code == kernel_input::RelativeAxes::X as u16 {
-        return Some(peridot::NativeAnalogInput::MouseX);
-    }
-    if code == kernel_input::RelativeAxes::Y as u16 {
-        return Some(peridot::NativeAnalogInput::MouseY);
-    }
-    if code == kernel_input::RelativeAxes::Wheel as u16 {
-        return Some(peridot::NativeAnalogInput::ScrollWheel);
-    }
-
-    map_input_rel(code)
-}
-fn map_input_rel(code: u16) -> Option<peridot::NativeAnalogInput> {
-    if code == kernel_input::RelativeAxes::X as u16 {
-        return Some(peridot::NativeAnalogInput::StickX(0));
-    }
-    if code == kernel_input::RelativeAxes::RX as u16 {
-        return Some(peridot::NativeAnalogInput::StickX(1));
-    }
-    if code == kernel_input::RelativeAxes::Y as u16 {
-        return Some(peridot::NativeAnalogInput::StickY(0));
-    }
-    if code == kernel_input::RelativeAxes::RY as u16 {
-        return Some(peridot::NativeAnalogInput::StickY(1));
-    }
-    if code == kernel_input::RelativeAxes::Z as u16 {
-        return Some(peridot::NativeAnalogInput::StickZ(0));
-    }
-    if code == kernel_input::RelativeAxes::RZ as u16 {
-        return Some(peridot::NativeAnalogInput::StickZ(1));
-    }
-
-    None
-}
-fn map_mouse_input_abs(code: u16) -> Option<peridot::NativeAnalogInput> {
-    if code == kernel_input::AbsoluteAxes::X as u16 {
-        return Some(peridot::NativeAnalogInput::MouseX);
-    }
-    if code == kernel_input::AbsoluteAxes::Y as u16 {
-        return Some(peridot::NativeAnalogInput::MouseY);
-    }
-    if code == kernel_input::AbsoluteAxes::Wheel as u16 {
-        return Some(peridot::NativeAnalogInput::ScrollWheel);
-    }
-
-    map_input_abs(code)
-}
-fn map_input_abs(code: u16) -> Option<peridot::NativeAnalogInput> {
-    if code == kernel_input::AbsoluteAxes::X as u16 {
-        return Some(peridot::NativeAnalogInput::StickX(0));
-    }
-    if code == kernel_input::AbsoluteAxes::RX as u16 {
-        return Some(peridot::NativeAnalogInput::StickX(1));
-    }
-    if code == kernel_input::AbsoluteAxes::Y as u16 {
-        return Some(peridot::NativeAnalogInput::StickY(0));
-    }
-    if code == kernel_input::AbsoluteAxes::RY as u16 {
-        return Some(peridot::NativeAnalogInput::StickY(1));
-    }
-    if code == kernel_input::AbsoluteAxes::Z as u16 {
-        return Some(peridot::NativeAnalogInput::StickZ(0));
-    }
-    if code == kernel_input::AbsoluteAxes::RZ as u16 {
-        return Some(peridot::NativeAnalogInput::StickZ(1));
-    }
-
-    None
 }
